requirements.txt
python-telegram-bot
requests
beautifulsoup4
python-dotenv
groq
schedule
